Fix error handling in init.configure
real_host() has an exception handler for ValueError, which will exit
with a helpful error message. But the exception assumes that
ValueError has a message field, which it doesn't, and the attribute
error that results swallows the actual error.

This just prints the string value of the exception instead which
provides a useful error.