Temporary: Remove color diagnostics options for compiler
color_cflags() doesn't quite work on Haiku at the moment, since it
assumes that either clang or gcc will be used. But in some cases clang
is running gcc to link, so the options get mixed up.

Just disabling this for now until I figure out a better way to fix
this.